add score display connected to hero
