+++ static/game/hero.js
@@ -6,6 +6,7 @@ import { Particle, EffectEnum } from "./particle.js";
 import { PowerUp } from "./powerup.js";
 import { Creature } from "./creature.js";
 import { playSound } from "../modules/sound.js";
+import { ScoreDisplay } from "./scoredisplay.js";
 
 const DEFAULT_SIZE = 50;
 
@@ -13,11 +14,12 @@ const DEFAULT_SIZE = 50;
 export const CHEAT_RADIUS = 16;
 
 export class Hero extends Creature {
+  /** @type {ScoreDisplay | undefined} */
+  scoreDisplay;
   drag = 0.1; // movement deceleration
   eyeDirection = new Vector(0, 1);
   invincibilityFrames = 0;
   invincibilityFramesMax = 100;
-  score = 0;
 
   /**
    * @param startingPos {Vector} the starting position of this Hero
@@ -159,6 +161,6 @@ export class Hero extends Creature {
    * @param {number} amt number of points to add
    */
   addPoints(amt) {
-    this.score += amt;
+    if (this.scoreDisplay !== undefined) this.scoreDisplay.addPoints(amt);
   }
 }
